Fix confusion between clearDisplay and clearMatrix
otherwise it won't compile with a call to clearMatrix() line 42 without defining the function. Added a clearDisplay function which clears the matrix then displays the cleared matrix. See if it is needed to be able to clear the display while keeping the matrix untouched. Assumed here that it's not needed.